// eventorganizer/frontend/event-frontend/src/app/core/model/User.ts
export interface User {
    id?: number; // Optional, as it's generated by the backend
    username: string; // User's login name
    password: string; // User's hashed password
    role: Role; // User's role (e.g., ADMIN, USER, etc.)
  }


  export enum Role{
    ADMIN='ADMIN',
    USER='USER',
    VENDOR='VENDOR',
  }